pcs commands on host: manila-share
This implements the creation of the manila-share bundle on the host.
The testing protocol used is documented in the depends-on.

The reason for adding a post_update task is that during a minor update
the deployment tasks are not run during the node update procedure but
only during the final converge. So we ran the role again there to make
sure that any config change will trigger a restart during the minor
update, so the disruption is only local to the single node being
updated. If we did not do this a final converge could potentially
trigger a global restart of HA bundles which would be quite disruptive.

Depends-On: Iaa7e89f0d25221c2a6ef0b81eb88a6f496f01696

Change-Id: If5730dba4973a5927c84565b6a65398ea1d7072f
Related-Bug: #1863442
(cherry picked from commit a1bee391361cde4345c2b78770865dab0278785e)