pcs commands on host: rabbitmq
This implements the creation of the rabbitmq bundle on the host.
The testing protocol used is documented in the depends-on.

The reason for adding a post_update task is that during a minor update
the deployment tasks are not run during the node update procedure but
only during the final converge. So we ran the role again there to make
sure that any config change will trigger a restart during the minor
update, so the disruption is only local to the single node being
updated. If we did not do this a final converge could potentially
trigger a global restart of HA bundles which would be quite disruptive.

NB: The init_bundle now has become the wait_bundle
as it just waits for rabbitmq to be up and functional

Related-Bug; #1863442

Depends-On: Iaa7e89f0d25221c2a6ef0b81eb88a6f496f01696
Change-Id: I853bcf354f64ef88cec9e68ad5c123e4af786de3