Add a 600s timeout when creating enable-ssh-admin workflow
If enable-ssh-admin.sh workflow fails, the script was looping forever,
because it only checks if the workflow is SUCCESSFUL.

Now, we have a timeout of 500 seconds and when it reaches it, it'll stop
and print a debug message.

Note: we're using ENABLE_SSH_ADMIN_TIMEOUT variable which can be
overriden when calling the script; the default value was taken from
tripleoclient/constants.

Change-Id: Ie183321bc23164745237d7036885f6e6dcfef859
Closes-Bug: #1876646
(cherry picked from commit b3127ac3f17d61e8bf6cfd43b05a4a2f31f7939f)