Neutron ML2/OVS: add support to enable of IGMP snooping
Neutron ML2/OVS backend allows now to enable support for IGMP
snooping in the br-int bridge. This can be done through
"igmp_snooping_enable" config option which can be set for
neutron-ovs-agent.
Patch [1] added support for it in puppet-neutron module.
In patch [2] support to enable this option for ML2/OVN backend
was added but we missed same change for ML2/OVS. Now we will
have it too.

[1] https://review.opendev.org/#/c/707367/
[2] https://review.opendev.org/#/c/714462/

Change-Id: I195410768d2d6a219bdda88a74fed3c29e1113a9