Expose grafana, dashboard and mgr port parameters
ceph-ansible computes facts to run ceph services on the proper
ip addresses, using default ports.
Due to "how ansible gather the instance facts" this can result
in an issue when haproxy services are included in the overcloud
deployment, because in a standalone environment, frontend and
backend ip addresses can be the same.
The purpose of this patch is to expose the port parameter of the
involved ceph services, making them configurable.
By doing this, we should be able to properly customize haproxy
with the default config provided by puppet manifests, customizing
the backend server address/ports in the job definitions.

Change-Id: I5b3b4b96a60c75826a24019932b368a9d2413385