Execute kvm-setup inside nova_libvirt container
On POWER8 (and some others), kvm requires some additional setup before
VMs can be successfully run. There is a script to perform this setup,
but it is provided as a systemd service -- we don't have systemd inside
containers, so let's include the invocation of the script in the
definition of the container.

Change-Id: I1e7208e4bd1c7f0c87354b6f6963c0d2df86cc02
Closes-Bug: #1878048