Configure valid_exit_code for startup containers
For the containers managed during the deploy tasks; they can also fail
after being run (e.g. nova db sync); and right now we don't catch it,
and keep the deployment moving on.

We catch the errors with puppet containers but not for startup
containers.
This patch is a first step in that direction where we only accept
containers that have exited to return 0 as a valid return code.

A patch in tripleo-ansible will be made to check these containers which
must have "command" in their configs.

Change-Id: I43e42df53b10fc99ca8e0fd8d7a30768e895e91f
Related-Bug: #1878074