Add option to not install ipa client packages
In RHEL 8, ansible-freeipa uses RHEL8 modules to determine which
IPA client packages to install.  If the system is not registered,
this fails because the RHEL 8 repos and groups are not available.

Fortunately, the IPA client packages are already in the ovecloud
image.  We add here an option to tell ansible-freeipa not to check
for and install any packages to the overcloud nodes to allow the
client registration to succeed..

Change-Id: I702bc4e5ea44af2e7f6247e5ad85ec41cfb17406
(cherry picked from commit bd132e20d4918936000e7d93655ddb582aa04bc7)