Use /32 netmask for VIPs
Prior to commit c712355e4bae4ef2fc1b83e5603c0364dbd50a78
KeepaliveD created the VIP addresses. KeepaliveD created
the VIPs with /32 netmask, when moving the VIPs to the
DeployedServerPortMap and adding them to the br-ctlplane
interface the netmask of the ctlplane subnet was used
(typically /24). The result is a routing table that
potentially uses the incorrect device for traffic when
the public VIP is not on in the ctlplane subnet.

This change hard-codes the netmask for the VIP addresses
to /32.

blueprint replace-keepalived-undercloud
Closes-Bug: #1878101
Change-Id: I873e925d2250677f25b9ae51ed0b87bd1b8e6b32