Change the :Z mount flag to :z
Leaving this with the capital Z limits access to this file
to the other novajoin container. This can cause the other
container to malfunction. At the very best, the other
container will restart, relabeling the file with it's
settings, which has the proper value of lower z.

Change-Id: Iebcfb9509c58619f9d68d3036775f07683a0fa42
(cherry picked from commit a6fa6027275bd650390cd538abffd7f06da24756)