[TRAIN ONLY] Introduce hybrid state for nova compute
Queens runs on CentOS7/RHEL8. Train runs on CentOS7 upstream,
but only RHEL8 downstream.

It is vastly impractical for a large compute environment to
upgrade the entire environment in one go. Typically there is
a need to do so in batches. Part of the process requires
ensuring that the compute service from Train is running on
the RHEL7 hypervisor so that we can live-migrate instances
between the RHEL7 hypervisor and a RHEL8 hypervisor.

This allows us to evacuate a hypervisor on RHEL7, upgrade or
replace it, then migrate instances on to it and continue
rolling to the next compute host.

This patch introduces this 'hybrid' state to make this
possible.

Change-Id: Ice37f2773d35e428a57006586397cc4cfaa99164