Fix sending SIGTERM to the sidecar containers
This is "repeat" of the Nate's fix done for stable/train branch in
[1].
This patch had got one issue - it missed passing "15" signal to the
signal_container() function and due to that e.g. keepalived sidecar
container wasn't never killed when router was deleted.

Patch [1] was recently removed with patch [2] so now I need to get it
back in the kill-script.

New version of the kill-script is in tripleo-ansible in the
tripleo-systemd-wrapper role where all is done fine and "15" signal is
passed properly to the signal_container() function.

[1] https://review.opendev.org/#/c/704463/
[2] https://review.opendev.org/#/c/724843/

Change-Id: I3f18922cf58bf8359f27aebf295bdb131bf2333a
Closes-bz: #1839071