MaxFailPercentage: default to 0
Switch the default MaxFailPercentage to 0, to match with the traditional
default.

We're currently working on making MaxFailPercentage working in the
TripleO playbooks (it doesn't work today); so this patch just fix the
default value, 15% was never applied in reality and we don't want to set
any default for backward compatibility. It'll be up to the deployers to
pick their number.

Change-Id: Ifea95c933e7a7d378ddf42050abda790e8631236
(cherry picked from commit 3c8f4a5afd60255e421f77633e8296c90e7e3711)