Check transfer data flag to skip pacemaker normal upgrade.
When the operating system upgrade is performed a
transfer data step must be executed (to allow the
db import once the operating system upgrades). During
this step the pacemaker cluster is stopped, so we can
create a new cluster with the newly OS upgraded node.
Therefore, as the pacemaker cluster is down we need
to skip some of the tasks which would be executed
during a normal upgrade (check pcs status, stop pcs and
start pcs) from the upgrade_tasks.

This patch removes the use of UpgradeLeappEnabled heat
parameter to identify this and uses the existence of
the flag file created during the transfer_data to skip
the normal pacemaker upgrade tasks by setting a new
fact, cluster_recreate, if this file exists.

Change-Id: Iba85e99f59258ce6ef4e05ccae737b9eeb6cfc57
(cherry picked from commit 5845c0a72aa68eb2dbbbdbd42fc1a6406aa92100)