Elevated privs are required to access files in home directory
Now that the config-download playbooks are located in the home
directory. We (tripleo-admin) needs elevated privs on the tasks
that require access to those file. This patch adds the required
become: true statements to those tasks.

Additionally, the current playbook looks for a specific ssh key,
if that key doesn't exist, it fails to ssh to the overcloud nodes.
To fix this, we can reference the ENV variable we're already
setting during the config-download deployment.

Change-Id: Ia97a5b0054bed697fc0390674fe4dba8317386a1
Closes-Bug: #1881420