Relax facts gathering plays on the overcloud
Tolerate failures on all the plays that run on the overcloud and which
potentially run with facts gathering only (no other tasks).

Since they can't apply a strategy, we want to let them fail without
error. In a later play, we have our tripleo strategies which will figure
out if we have reached the maximum of failures that was configured with
MaxFailPercentage.

Change-Id: I43652972494a9ed1f2b8b483fe19e685cbe4da1b