[TRAIN ONLY] Add FFU parameters in lifecycle env files.
The upgrade workflow has a set of environment files which are
used to set and unset parameters needed for the upgrade procedure.
This patch leverages this mechanism to minimize the amount of
parameters the user needs to add to be able to upgrade. This way
we reduce the possibilities of forgetting some of them by the user
but at the same time, allow the user to overwrite them if needed.

Change-Id: I8b02d4bb9d99d52b8fdf06dd43345814eba3272b