[TRAIN ONLY] Provide way to initialize Leapp
The Leapp utility can require enabling extra repositories and
it must be provided with data [1] to perform in-place upgrade
from RHEL7 to RHEL8. We allow passing thsese with parameters
LeappRepoInitCommand and LeappInitCommand. To make sure most
users will not have to pass LeappInitCommand we sync the files
from Undercloud. Lastly we default the UpgradeLeappCommandOptions
to have default OSP16 repositories enabled forcing users to
either change this value or properly subscribe.

Example:
LeappRepoInitCommand: |
  subscription-manager repos --enable rhel-7-server-rpms
  subscription-manager repos --enable rhel-7-server-extras-rpms
LeappInitCommand: |
  sudo curl -4ksSL -o /etc/leapp/files/pes-events.json https://example.com/pes-events.json
  sudo curl -4ksSL -o /etc/leapp/files/repomap.csv https://example.com/repomap.csv

[1] - https://access.redhat.com/articles/3664871

Resolves: rhbz#1845726
Change-Id: Ibd57468623717273c29d473d32a76f37ae02fa31