Don't set RABBITMQ_SERVER_ERL_ARGS
From the RabbitMQ docs[1]:

"RABBITMQ_SERVER_ERL_ARGS - Standard parameters for the erl command
used when invoking the RabbitMQ Server. This should be overridden for
debugging purposes only. Overriding this variable replaces the default
value."

We do not need any of the current values because of this commit from
2015 in rabbitmq 3.5.4:

https://github.com/rabbitmq/rabbitmq-server/commit/5c3c0e6f38792ca37289888ce39996e6baf4d846

In which the "+K true" and "-kernel inet_default_connect_options
"[{nodelay,true}]" options were hard-coded to the rabbitmq-server
script.  Meanwhile the "+P 1048576" configuration was moved to the
rabbitmq-env script under the SERVER_ERL_ARGS variable, which is the
defaults used when RABBITMQ_SERVER_ERL_ARGS is unset.

[1] https://www.rabbitmq.com/configure.html#supported-environment-variables

Let's also move the current RabbitAdditionalErlArgs to a new dedicated
override hiera key.

NB: To get this properly working we need a puppet-rabbitmq that contains
    https://github.com/voxpupuli/puppet-rabbitmq/pull/841
Depends-On: I3bf244a70538209773804eb85fae6be035c587f4

Closes-Bug: #1884922

Change-Id: I567839785a72813a382a00253562894e19eb6715