Convert roles section into tasks-include_role in deploy-steps.j2.
When executing the deployment_steps_playbook.yaml passing
a specific --tags, it looks like this play gets executed
even though the tag passed in --tags isn't common_roles.

This patch converts the roles: structure into a:
tasks:
  - include_role:

which is the prefered way since Ansible 2.4. This way, the
tags work properly and the execution of both roles is skipped
if the tag doesn't match common_roles.

Change-Id: I772ad486ca11525b8756a0b8cac7a5345373a5d3
Closes-Bug: #1885721