Update, avoid task skipping by directly importing step file.
During update we skip a lot tasks because we loop over the same update
step task file, changing the step variable, wasting time and
clobbering logs.

To solve the skipped tasks this, we now loop over the stepX file
generated for update and post update.

Expanding the playbook to import the tasks and setting the step
variable has another benefit.  It opens the possibility to use
"start-at-tasks" as everything is imported.  Using loop variable and
include_tasks prevented its usage.

Depends-On: https://review.opendev.org/740465
Change-Id: Ib32791d72410766a0b2e2330713120d15dc35f57