Set and then unset Ceph's noout flag before/after node is rebooted
The Ceph cluster needs not to mark down OSDs during the time when
the nodes are undercloud LEAPP upgrade because during that step
nodes will be rebooted.

This change implements setting and unsetting of the Ceph flags
(noout, norebalance, norecover and nobackfill) required to avoid
marking the OSDs down: the involved tasks are executed before node
(each) starts the LEAPP process and after the upgrade is finished and
the node rebooted.

This patch is going directly into train because it's trying to address
a problem that depends on the queens/train FFWD upgrade, while in
the main branch the upgrade_tasks list is still empty.

Change-Id: I1d21db4af894c096d929863daabd08673592370c