Add PermitRootLogin option in sshd_config before leapp.
One of the pre-requisites for leapp tool to be able to upgrade
is the existence of the PermitRootLogin option in the
sshd_config file. It doesn't mind which value it has, it needs
to be set. For that reason, this patch will set the value to
without-password (which is the default for the PermitRootLogin
field).

The selection of the option isn't really relevant, as during
the installation of the openstack services puppet will just
overwrite the sshd_config, therefore this is simply a hack
for leapp to let us upgrade.

Change-Id: Id647896fda6cc3fe7bf5eb5ba2de05111d174092
(cherry picked from commit d8e5ccb8cb3be3425c831c2e12407430b052cdde)
(cherry picked from commit 3e2a8d91cfd29b57a50df27845c63285de9853ee)