Add filestore to bluestore migration tags
Currently ceph-ansible assumes the cluster
(which is updated to ceph4) is using bluestore
just because it's the default on ceph-ansible.
This patch adds some missing tags to make sure
the variable contained in the heat stack are
properly regenerated and the fs2bs playbook can
be executed with the right variables.

Change-Id: I373636bbc9d296f41b63de76719320727d3f4fb9
(cherry picked from commit c9cd67124fb758e9023395e5ae4f8223111e2e51)