[TRAIN ONLY] Wait until DB is ready for neutron DB rename
On resource-constrainted systems, the database may take a little
extra time to come up. Currently this container fails if the
database is not immediately available.

In this patch we introduce a check to see if the database is
available, and a sleep/check cycle for up to two minutes before
timing out.

Resolves: rhbz#1851445
Change-Id: I50924f4f4d9606b6263e33dfe38b5bd37a6b56a5