Fix nCipher (aka thales) ansible role name
This patch fixes the role name used to run the ansible
role that configures nCipher (aka thales) HSM device when
Barbican is deployed with that backend.

The role was renamed from 'thales-hsm' to 'thales_hsm' [1]
but a couple of references were missed, which break the
deployment when using an nCipher device.

[1] https://review.opendev.org/#/c/724414

Change-Id: Ia9209cb4a781375577480c175126321515e7af7c