Remove /run from some services
redis(non-pcmk), nova-scheduler and swift-proxy do not need /run bind
mounted from the host.  As a matter of fact bind-mounting /run is
problematic due to a number of reasons (see LP#1883849 for more
background). In particular swift-proxy is the only swift container
(out of 9) that has /run bind-mounted.

These three services always had /run from the very beginning:
- redis -> Ie750caa34c6fa22ca6eae6834b9ca20e15d97f7f
- nova-scheduler -> I39436783409ed752b08619b07b0a0c592bce0456
- swift-proxy -> I2d96514fb7aa51dffe8fe293bc950e0e99df5e94

Tested this by applying this patch on a train deployment and
deployed an undercloud and an overcloud with it.
Verified that:
A) /run:/run is not present in the three containers
B) Deploy of UC and OC worked correctly
C) Tempest still works
D) Restarting the swift_proxy and nova_scheduler works correctly
E) Reboot the overcloud worked okay and tempest still works after the
   full overcloud reboot
F) Ran a minor UC update
G) Ran a minor update on all nodes and tempest still worked
H) Ran a redeploy on all nodes and tempest still worked

NB: I did not investigate other containers that bind mount /run
because they 1) seem to need it and 2) had no means to do proper
testing.

NB2: Note that while once we rebuild containers with
I81e5b7abf4571fece13a029e25911e9e4dece673 this change here is not
strictly needed for the LP bug, but is a nice cleanup nonetheless.
So this is to be backported only if rebuilding containers is
a problematic/costly move

Change-Id: Ic1a892a7f78a54b5e149f5ce52cb9db68ebc9529
Related-Bug: #1883849