Consolidate the barbican-api client configurations
Replace the cinder, glance and nova service_config_settings in the
BarbicanApi template with a reference to the settings in the
BarbicanClient template. This consolidates the settings in one place.

Update the BarbicanClient's service_config_settings so it handles
all cinder services that access barbican (c-api, c-volume, c-backup).
This change takes advantage of a recent enhancement in puppet-cinder
(see https://review.opendev.org/739126).

Depends-On: Ie3c95da2c0dab83e3c4b7e10f8a3531301692da5
Change-Id: I42b7c4a2add1dc25083c4c0e8a162ca4a3880e2a