Remove /var/lib/config-data context task
The task should not be needed since we set the context in
openstack-selinux:

https://github.com/redhat-openstack/openstack-selinux/blob/master/local_settings.sh.in#L89
fcontext -N -$1 -t container_file_t \"$LOCALSTATEDIR/lib/config-data(/.*)?\"

This task is very expensive at scale and we should avoid these king of
things.

svirt_sandbox_file_t is an alias to container_file_t, so let's use
container_file_t directly to be consistent and not creating confusion.

This patch shouldn't be backported to Train because in Train we'll have
this task not removed but moved to run only under FFU context.

Change-Id: I973a49aec1a263e2a1ec3e5199803c86d6f77b35
Closes-Bug: #1887721