Use a single task for fact gathering
We don't need to do different plays for fact gathering as we can simply
include both the source/remote hosts in a single pattern.

For the record, this was added because we were using local connections
for the undercloud when running under mistral however this is no longer
the case. Additionally I believe we stopped doing that later as well but
this may not be able to be backported to run under mistral. Original
context https://review.opendev.org/#/c/518323/

Change-Id: I4298855fe71b1035df81d5ce9c3cef2e3c96dbb6