pcmk_remote FFU support for Instance HA
The idea here is to set an override key like the following
during the upgrade:
hiera -c /etc/puppet/hiera.yaml pacemaker_remote_short_node_names_override
["compute-0"]

So that puppet-tripleo can detect the key and act accordingly knowning
that it is being upgraded (main reason is that the authkey management
for remotes needs to be special-cased in puppet).
Once the upgrade is completed we remove the key in post_upgrade_tasks.
Tested this on a number of runs and confirmed that:
- The IHA FFU completed correctly on controlplane + computes
- Tempest still works after the FFU
- The override keys are correctly removed at the end of the upgrade
  of each compute node

Closes-Bug: #1888398

Change-Id: I8bc42fb758a333adc9cd65602b44fabee6fc4041
(cherry picked from commit 2f460accb9800b2232c7cf783907b78bb94d98ef)