[queens-only] Unmount cinder NFS shares during upgrade
Unmount any cinder NFS shares on the host when upgrading cinder from
running on the host to running in a kolla container. This ensures
there are no active NFS mounts in cinder's service directory prior to
launching the container. Kolla recusively changes the ownership of
/var/lib/cinder, so removing the mounts prevents kolla from inadvertently
changing the ownership of files on the NFS storage backends.

Closes-Bug: #1887435
Change-Id: I5f0cedb808c6371caee6918da5b8bef743971221