Generated passthrough_whitelist shall use all the user_configs fields
Earlier, the derived pci passthrough whitelist includes the address,
vendor and product details only. This is not sufficient for nova to
allocate the VFs to the guests. Now, all the fields of the user_config
shall be used in the derived passthrough_whitelist. Only either of 'address'
or 'devname' shall be provided in passthrough_whitelist, and since the
'address' fields are must have to specify the induvidual VF's, the
'devname' is removed if present.

Change-Id: I5a337a67893241e1443ca2c3721f4c9dd78b95c8