Stop using a conditional for role tasks
Currently this task definition is causing the following warning to be
thrown at execution time:

    [WARNING]: conditional statements should not include jinja2 templating
    delimiters such as {{ }} or {% %}. Found: '{{ playbook_dir }}/{{
    _task_file_path }}' is exists

There is no reason to be using this conditional with a variable since
this file path is known when we create the playbook. Additionally we
don't need to use the playbook_dir because we're not using it in the
include_tasks. This allows us to specifically express the file that
we're using for these tasks inclusions rather than relying on a
conditional that is not recommended.

Change-Id: I368e99d2384469ca166e2e2b00e8621b7afe72db