Add openvswitch special treatment to update too.
Openvswitch package update can be disruptive during update too.  So we
factor out the code used for upgrade to re-use for update and
undercloud update.

We also fix a issue where yum/yumdownloader is not available anymore.
This cause the special treatment to not happen anymore what not matter
what.

Instead of duplicating the ansible sequence, this has been ported to a
ansible module that replicate the action defined in the templates.

We also make sure to not try to restart the openvswitch service if
nothing has been changed during the ovs upgrade task.  This cover two
cases.  First, it's an optimization, if nothing has been done, we
don't need to check the service.  Second, and most importantly, if ovs
isn't installed (this can happen on some minimal installation) then we
don't try to restart the non-existent service.

We keep the tasks at step 3 that make sure again that the openvswitch
service is started.  This is a failsafe and as such we ignore any
error there.  This also support the overcloud-minimal type of
installation.

Depends-On: https://review.opendev.org/742968
Change-Id: I4fde50e353ab7b2e6b03d963f585f6abe35ca32d
Closes-Bug: #1888651
Related-Bug: #1888821