Use appropriate allocation pools for StorageNFS
The StorageNFS network was defined in network_data_ganesha.yaml
using allocation ranges that mirrored those of the other isolated
networks.  But TripleO and the undercloud only need to use a small
number of addresses for overcloud deployment -- in the default
three-controller case, one for the regular StorageNFS interface on
each ControllerNfs role node, and one for the VIP on which the NFS
service is offered.  The bulk of the addresses in the StorageNFS CIDR
should be left out of the allocation range defined in network_data
so that they can be used by the allocation pool for the overcloud
Netron StorageNFS provider network's subnet without danger of
overlap.

This change uses a CIDR with a shorter prefix so that there will
be sufficient IPs left over after the undercloud's TripleO allocation
pool to deploy almost 4000 clients on the Neutron StorageNFS provider
network's subnet.

This commit also includes some minor changes to synchronize
network_data_ganesha.yaml with network_data.yaml since the former
was derived from the latter and should differ from it only by
inclustion of the StorageNFS network.

Closes-bug: #1889682

Change-Id: Ibb50dad42ec3dc154cd27ae9094a9be5d0a2dd28