[Train Only] Add CephAnsibleWarning into upgrade lifecycle environment files.
During the whole FFU, the ceph-ansible 3 package is left in the Undercloud
getting upgraded once the upgrade converge steps finishes. One of the
consequences of leaving ceph-3 running in a RHEL8 system is that the
repository gets lost. Therefore, the ceph-ansible 3 From-repo field is lost,
which is the value checked by ceph-ansible validations.
This patch sets to false the CephAnsibleWrning heat parameter to skip the
ceph-ansible repo validation and sets it back to true during the converge
step, so the validation will be in order when we will run the ceph upgrade
step.

Change-Id: Ie02a3a733f8bce8429c7e693f4e2766b289c66ab