[TRAIN-ONLY] Set the right container_client when set/unset noout
During FFU, when leapp execution is reached, ceph nodes should set
noout, norebalance and a few other options to avoid data rebalancing
during the reboot process.
After the nodes have been upgraded, we should unset these options.
Due to the fact these commands are run on different stages,
container_cli is different and should be computed looking at the current
status of the controller nodes.
The purpose of this patch is to fix the container_client used to run
commands against the Ceph cluster.

Change-Id: Ib10b6ad50e1052fe69640b39a173e856a426d607