Use sensible defaults for novnc TLS minimum versions and ciphers
We recently introduced a change that allowed operators to pass novnc TLS
cipher parameters to puppet-nova:

  https://review.opendev.org/#/c/723920/10

Unfortunately, the default values for NovaVNCProxySSLCiphers and
NovaVNCProxySSLMinimumVersion conflict with puppet-nova and causes TLS-e
deployments to fail with the following error during the overcloud
deployment:

  /var/log/containers/nova/nova-novncproxy.log:2020-08-03 04:45:41.120 8
  ERROR nova oslo_config.cfg.ConfigFileValueError: Value for option
  ssl_minimum_version from LocationInfo(location=<Locations.user: (4,
  True)>, detail='/etc/nova/nova.conf') is not valid: Valid values are
  [default, tlsv1_1, tlsv1_2, tlsv1_3], but found ''

This is because the values don't match what puppet-nova is expecting and
it causes the containers to fail.

This commit attempts to add some resonable defaults in THT that align
more closely with the puppet-nova defaults. It also only sets the
ciphers if they're set by the end user.

Change-Id: I2663bc9154846cc4642c3a030be0c57df4f25e1b